C++ set to 20; style changes

+++ src/graph/edge.h
@@ -7,13 +7,13 @@ class Edge
 {
 public:
 
-    Edge(char src, char dst, int weight) : src(src), dst(dst), weight(weight) {}
+    Edge(char src, char dst, size_t weight) : src(src), dst(dst), weight(weight) {}
 
     bool operator==(const Edge& edge);
 
     const char src;
     const char dst;
-    const int weight;
+    const size_t weight;
 };
 
 #endif
